requirements.txt
tensorflow-macos==2.16.2
tensorflow-metal==1.2.0
numpy
pandas
scikit-learn
xgboost==1.6.1

